Refactored PostPage for easier testing.
