+++ lib/view/widgets/user_details_with_follow.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:leaf/model/user_model.dart';
+import 'package:leaf/view/widgets/user_details.dart';
+
+class UserDetailsWithFollow extends StatelessWidget {
+  final UserModel userData;
+
+  const UserDetailsWithFollow({Key key, @required this.userData}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.center,
+      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+      children: <Widget>[
+        Expanded(flex: 2, child: UserDetails(userData: userData)),
+        Expanded(
+          flex: 1,
+          child: Container(
+            alignment: Alignment.centerRight,
+            child: IconButton(
+              icon: Icon(Icons.group_add),
+              onPressed: () {},
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
